requirements.txt
nltk==3.3
python-twitter==3.4.2